pyproject.toml
[project]
name = "girdermedviewer"
version = "1.0.0"
description = "Browser and Viewer of Girder database"
authors = [
    {name = "Kitware Europe"},
]
dependencies = [
    "girder-client==3.2.11",
    "trame==3.12.0",
    "trame-gwc==0.2.0",
    "trame-server==3.10.0",
    "trame-vuetify==3.2.0",
    "trame-vtk==2.10.2",
    "py-undo-stack==1.1.0",
    "vtk==9.5.2"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache Software License"}
keywords = ["Python", "Interactive", "Web", "Application", "Framework"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
app = [
    "pywebview",
]
jupyter = [
    "jupyterlab",
]
dev = [
    "pre-commit",
    "ruff",
    "pytest",
]
dicom = [
    "dicom-exporter==1.0.0",
]

[project.scripts]
girdermedviewer-cli = "girdermedviewer.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["girdermedviewer"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.semantic_release]
version_variables = [
    "girdermedviewer/__init__.py:__version__",
]
version_toml = [
    "pyproject.toml:project.version",
]
build_command = """
  python -m venv .venv
  source .venv/bin/activate
  pip install -U pip build
  python -m build .
"""

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[tool.ruff]

[tool.ruff.lint]
extend-select = [
    "ARG", # flake8-unused-arguments
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "EM", # flake8-errmsg
    "EXE", # flake8-executable
    "G", # flake8-logging-format
    "I", # isort
    "ICN", # flake8-import-conventions
    "NPY", # NumPy specific rules
    "PD", # pandas-vet
    "PGH", # pygrep-hooks
    "PIE", # flake8-pie
    "PL", # pylint
    "PT", # flake8-pytest-style
    "PTH", # flake8-use-pathlib
    "RET", # flake8-return
    "RUF", # Ruff-specific
    "SIM", # flake8-simplify
    "T20", # flake8-print
    "UP", # pyupgrade
    "YTT", # flake8-2020
]
ignore = [
    "G004",
    "EM101",
    "EM102",
    "PLR09", # Too many <...>
    "PLR2004", # Magic value used in comparison
    "ISC001", # Conflicts with formatter
]
isort.required-imports = []
